// frontend/src/api/generated/model/updateDataRoomCategoryDtoCategory.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export type UpdateDataRoomCategoryDtoCategory = typeof UpdateDataRoomCategoryDtoCategory[keyof typeof UpdateDataRoomCategoryDtoCategory];


export const UpdateDataRoomCategoryDtoCategory = {
  pitch_deck: 'pitch_deck',
  financial: 'financial',
  cap_table: 'cap_table',
  legal: 'legal',
  technical_product: 'technical_product',
  business_plan: 'business_plan',
  market_research: 'market_research',
  contract: 'contract',
  team_hr: 'team_hr',
  miscellaneous: 'miscellaneous',
} as const;